test: 拆分网络追踪测试结构
